Make gmediarender compile with version libupnp 1.6.7 ... 1.6.16.
With version 1.6.7 and above, the UPNP library maintainers made a questionable
API choice to register the webserver callbacks one-by-one, instead of having a
struct with the callbacks which is certainly a better and 'typical' C API
choice. They removed the UpnpVirtualDirCallbacks in the course of that.
Because it breaks code, it has been reverted in version 1.6.16
(see http://sourceforge.net/p/pupnp/bugs/29/ ), but we essentially have a
broken version between 1.6.7 ... 1.6.16.

Assuming that they will go on with this broken idea and eventually remove
the support for the VirtualDirCallbacks in new major versions, we now
register the callbacks separately when upnp-version >= 1.6.7