Retry when encountering UPnP socket initialization issues.
The UPnP socket can not bind to 0.0.0.0 so needs a stable interface
with a configured IP address to bind to.
If the OS has not initialized the network yet, this will fail, so
add a couple of rounds of retries to make this more robust.

This only showed up as an issue with recent distributions using
systemd. I suspect a bug in debian systemd network-online.target actually
not doing as advertised and claiming readiness when in fact DHCP might
still be ongoing.

Fixes #129