create new output_module element
The choice of the module is a part of the main part of the application.
To have the module selection tool into the ouput element force to have
all modules into the ouput. In this case a new module is only allowed
for gmrender application.
The goal is to reuse gmrender into other media application.
It is possible to dispatch gmrender between an application (main.c +
output_module.c) and a library (all other files).