remove some dependence to the glib
Glib is necessary only for gstreamer and the options parsing into the
main application.
gmediarenderer has a module loader and modules may not need glib. And in
this case glib is a big library to only parse the arguments.
Another step will be to allow to build gmediarenderer without glib.