add mpg123 as output module
mpg123 is a mpeg 1 & 2 layer 1,2 and 3 decoder library.
https://www.mpg123.de/index.shtml
mpg123 is a decoder only and it needs a sink driver.
For that sound_alsa is a sink for linux Alsa system. sound_module allows
to add other sink module for mpg123 as jack or pulseaudio.
sound_module is written like the output module except that it is not
possible to choice another module. The choice of the sink module has to
be done during the build configuration.

mpg123 allows to build a connected speaker with really small memory foot print.