output may load dynamicly the output library
By default gmrender hasn't any ouput's module, and should load at least
one.
Two way are availables:
 - use LD_PRELOAD to load the library before the application's launch:
$ LD_PRELOAD=/usr/local/lib/libgmrender_gst.so
/usr/local/bin/gmediarender
 - use the output option to set the name of the module to load:
$ /usr/local/bin/gmediarender -o gst

The second solution disallows the use of gstreamer's options on the
command line. This point must be corrected in the next commit.
This solution looks the library inside the LIBDIR directory and the name
must have the format of libgmrender_<shortname>.so

Because the core application doesn't use directly module API is
impossible the link this binary to a module library. The linker
disallows that.