first contribs with mpg123 module
This contribs is just an example how to create a new output module. It's
not stable.

output_mpg123.c is the real module
sound_alsa and sound_module manages the sound playback with alsa (a
version with tinyalsa may be contributed too in the future)
webclient allows to send GET request to the server and to receive the
audio file data.