split add_options feature between core and module
add_options feature was mixed inside main.c between the options for the
core application and the options for the modules.
To keep the use of glib API inside the module and to remove the glib
from main.c file, the feature has to be splited.
The consequence of this patch is that the options must be separated on
the command line with "--".
ex:
$ /usr/local/bin/gmediarender -o gst -- --gstout-videosink=qtvideosink